mcp/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cdmx-data"
version = "0.1.0"
description = "Hub de datos abiertos de la Ciudad de México"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "pandas>=2.0",
    "typer>=0.12",
    "rich>=13.0",
    "python-dateutil>=2.9",
]

[project.optional-dependencies]
geo = ["geopandas>=0.14", "shapely>=2.0"]
mcp = ["mcp>=1.0"]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-httpx>=0.30",
]

[project.scripts]
cdmx = "cdmx_data.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/cdmx_data"]

[tool.pytest.ini_options]
testpaths = ["tests"]
